requirements.txt
# Vercel deployment
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0

